Consolidated BelugaPlot2 features into BelugaPlot
Incorporated xlim, ylim, grid on features for matplotlib. User can specify a single color for a plot. Otherwise, the default color of blue is used for a single line plot and the jet colormap is used for line_series. Also reverted back to beluga dae code that works with the brachistochrone example.