Implemented custom plotting functions and removed interpolation default
Custom functions can be included in the plotting script. A hypersonic3DOF example is provided. The function must take in arguments that are valid problem variables (they are input as numpy arrays), and the numpy array returned from the function is plotted. Also, interpolation and mesh overwriting is disabled by default.